fix(cli): fix the "FileNotFoundError" from CLI "gas config" in windows
Root Cause:
The "HOMEPATH" in window environment variable is "\Users\<UserName>",
which is a relative path. It causes the config file cannot be found
when the CLI is not executed from "HOMEDRIVE" (C:\).

Solution:
Use "USERPROFILE" in window environment variable which is normally:
"C:\Users\<UserName>", which is a absolute path.