feat(dataset): use str() instead of ULID.str to transform ULID to str
If user use a "str" rather than "ULID" to initialize a Frame, the
"frame_id" will be a "str". Type "str" does not have ".str" property,
but function "str()" supports both "str" and "ULID". This makes
"Frame.dumps()" and "FusionSegmentClient.upload_frame()" work correctly
even if the "frame_id" is a "str"